+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v1.1.24
 - even more item rotations!
+- hopefully fix skinnedRenders
 
 ## v1.1.23
 - added some rotation values to the list